requirements.txt
apache-airflow==2.7.3
pyspark==3.4.2
unidecode==1.3.7
apache-airflow-providers-docker